Extract common Entry operations in an abstract class
